Fix treeview menù, fix carousel
