+++ design/designitalia/templates/openpa/full/parts/section_left.tpl
@@ -1,6 +1,6 @@
 {if or($openpa.control_menu.show_extra_menu, $show_left, is_set($extra_template))}
     <div class="extra">
-        {if or(is_set($extra_template), $openpa.control_menu.show_extra_menu)}            
+        {if or(is_set($extra_template), $openpa.control_menu.show_extra_menu)}
 			{if and(is_set($extra_template), $extra_template|ne(false()))}
                 {include uri=$extra_template}
             {else}
@@ -21,7 +21,7 @@
                             {/if}
                         </h2>
                     {/if}
-                    <ul class="Linklist Linklist--padded u-layout-prose Treeview js-fr-treeview u-text-r-xs">
+                    <ul class="Linklist Linklist--padded Treeview Treeview--default js-Treeview u-text-r-xs">
                         {foreach $tree_menu.children as $menu_item}
                             {include name=side_menu uri='design:menu/side_menu_item.tpl' menu_item=$menu_item current_node=$node recursion=1}
                         {/foreach}
