+++ design/designitalia/templates/page_footer.tpl
@@ -141,8 +141,8 @@
 
 
 <a href="#" title="torna all'inizio del contenuto" class="ScrollTop js-scrollTop js-scrollTo">
-    <i class="ScrollTop-icon Icon-arrow-up" aria-hidden="true"></i>
-    <span class="u-hiddenVisually">torna all'inizio del contenuto</span>
+  <i class="ScrollTop-icon Icon-collapse" aria-hidden="true"></i>
+  <span class="u-hiddenVisually">torna all'inizio del contenuto</span>
 </a>
 
 {undef}
