+++ design/designitalia/templates/page_footer_script.tpl
@@ -9,8 +9,10 @@
 <script src="{'javascript/vendor/polyfill.min.js'|ezdesign(no)}"></script>
 <![endif]-->
 
+<script src="{'javascript/plugins/owl-carousel/owl.carousel.min.js'|ezdesign(no)}"></script>
+
 {if ezini('TemplateSettings','DevelopmentMode')|eq('enabled')}
-  <script src="{'javascript/IWT.js'|ezdesign(no)}"></script>
+  <script src="{'javascript/IWT.min.js'|ezdesign(no)}"></script>
 {else}
   <script src="{'javascript/IWT.min.js'|ezdesign(no)}"></script>
 {/if}
@@ -18,7 +20,7 @@
 {if openpaini( 'Seo', 'GoogleAnalyticsAccountID', false() )}
 <script type="text/javascript">
   var _gaq = _gaq || [];
-  _gaq.push(['_setAccount', '{openpaini( 'Seo', 'GoogleAnalyticsAccountID' )}']);  
+  _gaq.push(['_setAccount', '{openpaini( 'Seo', 'GoogleAnalyticsAccountID' )}']);
   _gaq.push(['_setAllowLinker', true]);
   _gaq.push(['_trackPageview']);
   (function() {ldelim}
